requirements.txt
Django==5.2.4
djangorestframework==3.14.0
pymongo==4.6.0
geopy==2.4.0
python-decouple==3.8
django-cors-headers==4.3.1
drf-yasg==1.21.7
gunicorn==21.2.0 